Add files via upload
The "web-scraping-covid19-powerbi-Visualization" repository is a comprehensive resource that combines web scraping techniques and Power BI visualization capabilities to create dynamic and insightful COVID-19 dashboards. By leveraging Python libraries such as BeautifulSoup and Requests, users can extract up-to-date COVID-19 data from reputable sources across the web.

With the extracted data at hand, the repository provides pre-built Power BI templates and examples to enable users to visualize and analyze various aspects of the pandemic. These dashboards offer a range of interactive visualizations, including geographic heatmaps, time series charts, and trend analysis, allowing users to gain a deeper understanding of the global or regional impact of COVID-19.

The repository's focus on web scraping ensures that the data used for visualization is always current, enabling users to stay informed about the latest developments in the pandemic. The integration with Power BI empowers users to explore the data visually, uncover patterns, and communicate insights effectively.

Whether you are a data analyst, a researcher, or an organization seeking real-time insights on COVID-19, the "web-scraping-covid19-powerbi-Visualization" repository provides a valuable toolkit for extracting and visualizing data in a user-friendly and impactful manner. Stay up to date, track trends, and make informed decisions with this powerful combination of web scraping and Power BI visualization.


The data used in the "web-scraping-covid19-powerbi-Visualization" repository is sourced from the WorldCovidMeter website.